// types/config.ts
export type MonitorTarget = {
  id: string
  name: string
  method: string
  target: string
  enabled?: boolean
  statusPageLink?: string
  expectedCodes?: number[]
  timeout?: number
  headers?: { [key: string]: string | number }
  body?: string
  responseKeyword?: string
  responseForbiddenKeyword?: string
  expectedJson?: { [key: string]: string | number | boolean | null }
  checkProxy?: string
  checkProxyFallback?: boolean
}

export type WorkerConfig<TEnv = Env> = {
  stateWriteCooldownMinutes?: number
  notification?: Notification
  callbacks?: Callbacks<TEnv>
}

export type Notification = {
  webhook?: WebhookConfig
  timeZone?: string
  gracePeriod?: number
  skipNotificationIds?: string[]
  skipErrorChangeNotification?: boolean
}

export type SingleWebhook = {
  url: string
  method?: 'GET' | 'POST' | 'PUT' | 'PATCH'
  headers?: { [key: string]: string | number }
  payloadType: 'param' | 'json' | 'x-www-form-urlencoded'
  payload: any
  timeout?: number
}

export type WebhookConfig = SingleWebhook | SingleWebhook[]